feat: Move FirebaseAuthRepository to data module (Phase 22)
Move FirebaseAuthRepository from androidApp to data/src/commonMain/.
Change dependency from AppLoggerRepository (Android) to AppLoggerDataSource
(shared). Keep RC_ONE_TAP_SIGN_IN in androidApp (Android sign-in specific).
