+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/auth/AuthConstants.kt
@@ -1,0 +1,4 @@
+package com.chriscartland.garage.auth
+
+/** Request code for Google One-Tap sign-in (Android-specific). */
+const val RC_ONE_TAP_SIGN_IN = 1
